third_party/WebKit/LayoutTests/imported/wpt/web-animations/interfaces/AnimationEffectTiming/easing-expected.txt

This is a testharness.js-based test.
FAIL step-start function assert_equals: expected "steps(1, start)" but got "step-start"
FAIL steps(1, start) function assert_equals: expected "steps(1, start)" but got "step-start"
PASS steps(2, start) function 
FAIL step-end function assert_equals: expected "steps(1)" but got "step-end"
FAIL steps(1) function assert_equals: expected "steps(1)" but got "step-end"
FAIL steps(1, end) function assert_equals: expected "steps(1)" but got "step-end"
FAIL steps(2, end) function assert_equals: expected "steps(2)" but got "steps(2, end)"
PASS linear function 
PASS ease function 
PASS ease-in function 
PASS ease-in-out function 
PASS ease-out function 
PASS easing function which produces values greater than 1 
PASS Invalid effect easing value test: '' 
PASS Invalid effect easing value test: 'test' 
PASS Invalid effect easing value test: 'cubic-bezier(1.1, 0, 1, 1)' 
PASS Invalid effect easing value test: 'cubic-bezier(0, 0, 1.1, 1)' 
PASS Invalid effect easing value test: 'cubic-bezier(-0.1, 0, 1, 1)' 
PASS Invalid effect easing value test: 'cubic-bezier(0, 0, -0.1, 1)' 
PASS Invalid effect easing value test: 'steps(-1, start)' 
PASS Invalid effect easing value test: 'steps(0.1, start)' 
PASS Change the easing while the animation is running 
Harness: the test ran to completion.
